Refactor live scoreboard into its own ES6 module
