+++ src/resources/assets/js/live_scoreboard/GameField.jsx
