+++ src/resources/assets/js/live_scoreboard/GameRow.jsx
