+++ src/resources/assets/js/live_scoreboard/LiveScoreboard.jsx
