+++ src/resources/assets/js/live_scoreboard/index.js
@@ -1,0 +1,3 @@
+import LiveScoreboard from './LiveScoreboard'
+
+export { LiveScoreboard }
